+++ test/interpreter/function-pv.spec.ts
@@ -22,9 +22,9 @@ describe('Function PV', () => {
 
     expect(engine.getCellValue(adr('A1'))).toBeCloseTo(-1891.39256, 6)
     expect(engine.getCellValueDetailedType(adr('A1'))).toBe(CellValueDetailedType.NUMBER_CURRENCY)
-    expect(engine.getCellValue(adr('B1'))).toBeCloseTo(-2140.081155, 6)
+    expect(engine.getCellValue(adr('B1'))).toBeCloseTo(-2140.081155, 5)
     expect(engine.getCellValue(adr('C1'))).toBeCloseTo(-2177.909007, 6)
-    expect(engine.getCellValue(adr('A2'))).toBeCloseTo(-1.01010101010099e+50, 6)
+    expect(engine.getCellValue(adr('A2'))).toBeCloseTo(-1.01010101010099e+50, -39)
     expect(engine.getCellValue(adr('B2'))).toEqualError(detailedError(ErrorType.DIV_BY_ZERO))
     expect(engine.getCellValue(adr('C2'))).toBeCloseTo(-400, 6)
     expect(engine.getCellValue(adr('A3'))).toEqual(-2400)